Collect and dump footnotes on footnote-list block and at end of post.
In this patch we're starting to collect footnote references on every
block and then flushing them out to a list whenever we reach a
footnote list block. If multiple blocks exist we'll end up with
multiple footnote lists. If we reach the end of the post and still
have footnotes not listed we will flush those out too.

This is some exploratory work to enable better in-editor experience
previewing footnotes and moving them to another place, presumably
such as an `<aside>` in a template.

 - Also adds `title` to `<sup>` so hover reveals the foonote.